Renamed notifications to activity
