+++ apps/admin-x-activitypub/src/components/navigation/MainNavigation.tsx
@@ -14,7 +14,7 @@ const MainNavigation: React.FC<MainNavigationProps> = ({}) => {
             <div className='col-[2/3] flex items-center justify-center gap-9'>
                 <Button icon='home' iconColorClass={mainRoute === '' ? 'text-black' : 'text-grey-500'} iconSize={18} unstyled onClick={() => updateRoute('')} />
                 <Button icon='magnifying-glass' iconColorClass={mainRoute === 'search' ? 'text-black' : 'text-grey-500'} iconSize={18} unstyled onClick={() => updateRoute('search')} />
-                <Button icon='bell' iconColorClass={mainRoute === 'notifications' ? 'text-black' : 'text-grey-500'} iconSize={18} unstyled onClick={() => updateRoute('notifications')} />
+                <Button icon='bell' iconColorClass={mainRoute === 'activity' ? 'text-black' : 'text-grey-500'} iconSize={18} unstyled onClick={() => updateRoute('activity')} />
                 <Button icon='user' iconColorClass={mainRoute === 'profile' ? 'text-black' : 'text-grey-500'} iconSize={18} unstyled onClick={() => updateRoute('profile')} />
             </div>
             <div className='col-[3/4] flex items-center justify-end px-8'>
